Improve Select components
