+++ src/components/PersonSelect.tsx
@@ -1,16 +1,19 @@
 import { PersonSummary } from "apiClient/people";
 import { debounce } from "lodash";
 import { useMemo, useState } from "react";
-import Select from "react-select";
 
 import { usePeopleList } from "redux/api";
+
+import { Select } from "./Select";
 
 type Props = {
   value: PersonSummary;
   onChange: (person: PersonSummary | null) => void;
+  className?: string;
+  invalid?: boolean;
 };
 
-export function PersonSelect({ value, onChange }: Props) {
+export function PersonSelect({ value, onChange, className, invalid }: Props) {
   const [query, setInput] = useState<string>("");
   const debouncedSetInput = useMemo(() => debounce(setInput, 250), [setInput]);
   const { personList, isFetching } = usePeopleList({
@@ -25,12 +28,13 @@ export function PersonSelect({ value, onChange }: Props) {
 
   return (
     <Select
-      className="flex-grow-1"
+      className={`flex-grow-1 ${className ?? ""}`}
       options={options}
       value={value}
       onChange={onChange}
       onInputChange={debouncedSetInput}
       isLoading={isFetching}
+      invalid={invalid}
     />
   );
 }
