+++ src/pages/Gear/AddNewGear/AddNewGearForm.tsx
@@ -1,5 +1,4 @@
 import { useEffect } from "react";
-import Select from "react-select";
 import { useForm, FormProvider } from "react-hook-form";
 
 import { useGetGearTypesQuery } from "redux/api";
@@ -7,6 +6,7 @@ import { CreateGearArgs, GearType } from "apiClient/gear";
 import { LabeledInput } from "components/Inputs/LabeledInput";
 import { Link } from "react-router-dom";
 import { fmtAmount } from "lib/fmtNumber";
+import { Select } from "components/Select";
 
 type GearTypeOption = GearType & { value: string; label: string };
 
@@ -59,7 +59,7 @@ export function AddNewGearForm({
   }, [getFieldState, setValue, gearType]);
   const options =
     gearTypes?.map((gearType) => ({
-      value: gearType.id,
+      value: String(gearType.id),
       label: gearType.typeName,
       ...gearType,
     })) ?? [];
@@ -86,19 +86,11 @@ export function AddNewGearForm({
               <Select
                 isLoading={!gearTypes}
                 options={options}
-                className={`w-100 ${invalid ? "is-invalid" : ""}`}
-                styles={{
-                  control: (base, state) =>
-                    !invalid
-                      ? base
-                      : {
-                          ...base,
-                          ...invalidFormControlStyle,
-                        },
-                }}
+                className="w-100"
                 value={value}
                 onChange={onChange}
                 onBlur={onBlur}
+                invalid={invalid}
               />
             );
           }}
@@ -195,13 +187,3 @@ const exampleDescription = [
   "Ex: Includes: tent body, rain fly, footprint, 2 poles, stakes, carrying case.",
   "Tents must be set up and inspected when returned",
 ].join("\n");
-
-// This is copy-pasted from bootstrap, since we can't set the inner class of React-Select
-const invalidFormControlStyle = {
-  borderColor: "#dc3545",
-  paddingRight: "calc(1.5em + .75rem)",
-  backgroundImage: `url("data:image/svg+xml,%3csvg xmlns=%27http://www.w3.org/2000/svg%27 viewBox=%270 0 12 12%27 width=%2712%27 height=%2712%27 fill=%27none%27 stroke=%27%23dc3545%27%3e%3ccircle cx=%276%27 cy=%276%27 r=%274.5%27/%3e%3cpath stroke-linejoin=%27round%27 d=%27M5.8 3.6h.4L6 6.5z%27/%3e%3ccircle cx=%276%27 cy=%278.2%27 r=%27.6%27 fill=%27%23dc3545%27 stroke=%27none%27/%3e%3c/svg%3e")`,
-  backgroundRepeat: "no-repeat",
-  backgroundPosition: "right calc(.375em + .1875rem) center",
-  backgroundSize: "calc(.75em + .375rem) calc(.75em + .375rem)",
-};
